refactor: modularize code and remove comments
