+++ main.py
@@ -1,8 +1,8 @@
-from app import main
+from ui import main
 from oauth import main as flask_app
 import threading
 import requests
-from common import ShowWindowSignal
+from signals import ShowWindowSignal
 import socket
