+++ requirements.txt
@@ -5,4 +5,3 @@ keyring
 requests
 PySide6
 darkdetect
-aiohttp
